Simplified 'pub get' step, added make zip target.
Now make zip will create a new directory zip/ inside pub/ which only contains
the files necessary for inclusion in the packaged app.

Also the requirement for package:browser is removed, as the necessary code is
extracted to js/browser.js. The package:browser/dart.js part of the file has
been modified so that .dart files are replaced in the DOM by .precompiled.dart
files. Also the path is changed to be in js/ instead of lib/src/app.

The next patch will properly removed the duplicate files and fix the build
system so that the app/ directory will not be required.

R=tapted@chromium.org

Review URL: https://codereview.appspot.com/51210043