Added working directory bar.
This CL adds a button to the html allowing the working directory to be
changed after startup. The current working directory basename is shown
in a disabled text field next to it. Changing the working directory
persists the new entryID from chrome.fileSystem.retainEntry to local
storage.

R=tapted@chromium.org

Review URL: https://codereview.appspot.com/51260043